fix(install): corrections ShellCheck et code de sortie des tests PowerShell en CI
- install.sh : suppression de dossier protégée contre une variable vide (SC2115)
- test-install.sh : cd vérifié (SC2164)
- test-install.ps1 : exit 0 explicite, GitHub Actions terminant l'étape
  par exit $LASTEXITCODE

+++ install.sh
@@ -179,7 +179,7 @@ for item in "${COPY_ITEMS[@]}"; do
     # laisse l'ancienne version intacte.
     rm -rf "$CLAUDE_DIR/$item.new"
     cp -R "$SRC_DIR/$item" "$CLAUDE_DIR/$item.new"
-    rm -rf "$CLAUDE_DIR/$item"
+    rm -rf "${CLAUDE_DIR:?}/${item:?}"
     mv "$CLAUDE_DIR/$item.new" "$CLAUDE_DIR/$item"
     task_end DONE
   fi
